Responsive Functions
Many projects still use jquery lower than 1.7, therefore found it
necessary to update this plugin. It is very simple to use, and has a
separate code as well.

What was added?

OPTIONS
  - Responsive / / like minSlides and maxSlides the bxSlider 4!
                         / / Sets the displaySlideQty, depending on the
available area.

- MinSlides / / sets the minimum accepted displaySlideQty
- MaxSlides / / sets the minimum accepted displaySlideQty

METHODS
- ReloadSlider / / equal to the bxSlider 4, you may also pass an options
object
- CalculateDisplayQty / / Returns maximum amount of display que fits on
the slide
- AutoUpdateDisplayQty / / Update maximum amount of display que fits on
the slide, if you need to call without being on window resize